+++ app/src/main/java/com/moroom/android/SaveReviewWithLatLng.java
@@ -15,6 +15,7 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.Query;
 import com.google.firebase.database.ValueEventListener;
+import com.moroom.android.data.model.WrittenReviewData;
 
 import java.io.IOException;
 import java.util.ArrayList;
